pyproject.toml
[project]
name = "ss-test"
version = "0.1.0"
description = ""
authors = [
    {name = "MP"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm (>=1.61.20,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

